% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean Names}
\usage{
clean_names(
  df,
  by = c("adm1", "adm2"),
  code_col = "pcode",
  ref = NULL,
  fn_ref = NULL,
  fixes = NULL,
  fn_fixes = NULL,
  save_unmatched = ifelse(is.null(fn_fixes), FALSE, TRUE),
  require_all = TRUE,
  keep_raw = FALSE
)
}
\arguments{
\item{df}{\code{dataframe} Dataframe to be checked.}

\item{by}{\verb{chr vct} Column name(s) in \code{df} to be checked against \code{ref}. Default is to look for
columns called \code{adm1} and \code{adm2}.}

\item{code_col}{\code{chr} Column in \code{ref} to be used as the code column, for example for matching to
a \code{fixes} dataset. Default is \code{pcode}.}

\item{ref}{\code{dataframe} Dataframe of reference data. \strong{NOTE}, at least one of \code{ref} or \code{fn_ref}
\emph{must} be provided. If both are provided only \code{ref} is considered.}

\item{fn_ref}{\code{chr} Path to a set of reference data. \strong{NOTE}, at least one of \code{ref} or \code{fn_ref}
\emph{must} be provided. If both are provided only \code{ref} is considered.}

\item{fixes}{\code{dataframe} Dataframe of manual fixes. This dataframe \emph{must} contain the a code
column with the appropriate name (as specified by \code{code_col}).}

\item{fn_fixes}{\code{chr} Path to set of manual fixes data. This data \emph{must} contain the a code
column with the appropriate name (as specified by \code{code_col}).}

\item{save_unmatched}{\code{bool} If TRUE, unmatched values will be saved to \code{fn_fixes} for the user
to provide manual fixes. Default behavior is for this to happen whenever \code{fn_fixes} is not NULL
and not otherwise.}

\item{require_all}{\code{bool} If TRUE, \code{clean_names()} will return an error if it fails to find a
match for all values. Default is TRUE.}

\item{keep_raw}{\code{bool} If TRUE, the original unmatched data will be retained in the returned
output. The relevant columns will have "_raw" appended to their names. Default is FALSE.}
}
\description{
Convenience wrapper for a pipeline of name cleaning methods using the package
\href{https://epicentre-msf.github.io/hmatch/index.html}{hatch} for hierarchical fuzzy matching.
If hmatch is not installed, \code{clean_names()} can still take a manual fixes file and apply those
corrections to the dataset. \strong{Warning:} at present, this function will necessarily return a
\link{tidytable}.
todo: convert to method style to preserve var type between data.frame, tibble, and tidytable
}
