% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{drc}
\alias{drc}
\title{The World}
\format{
A \code{dataframe} with 147,851 observations of 4 fields:
\describe{
\item{prov}{Province Name}
\item{zs}{Health Zone Name}
\item{numsem}{ISO Week Number}
\item{MALADIE}{Maladie Name}
}
}
\usage{
drc
}
\description{
Countries of the World
}
\keyword{dataset}
