% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean_strings}
\alias{clean_strings}
\title{Clean Strings}
\usage{
clean_strings(x)
}
\arguments{
\item{x}{\verb{chr vct} Character vector to be cleaned.}
}
\description{
Cleans strings to squish unwanted spaces and translate non-ascii characters.
TODO: Optionally translate to lower, upper, or title case. Optionally replace spaces and hyphens
with underscores.
}
\examples{
 x <- c(' too    many spaces  ',
        '\u00c5ccents ar\u00e9 the w\u00f4rst')

}
